lua_types: included latest ultraschall documentation
resources/reaper-types.lua is now update, though I'm still struggling to
figure out how to get the correct path scheme to register the lua
package path.
Doing the registration is what allows files to require each other. If
the main.lua script is called with a relative path, lua can't find the
absolute location, but if I use absolute path, then I have to update the
requires and locations of files to be written to disk.
Will have to check how to reconcile this…

path update